+++ pipelex/core/pipes/pipe_output.py
@@ -2,7 +2,7 @@ from typing import TypeVar
 
 from pydantic import BaseModel, Field
 
-from pipelex.core.memory.working_memory import WorkingMemory
+from pipelex.core.memory.working_memory import DictWorkingMemory, WorkingMemory
 from pipelex.core.stuffs.html_content import HtmlContent
 from pipelex.core.stuffs.image_content import ImageContent
 from pipelex.core.stuffs.list_content import ListContent
@@ -13,6 +13,11 @@ from pipelex.core.stuffs.stuff_content import StuffContentType
 from pipelex.core.stuffs.text_and_images_content import TextAndImagesContent
 from pipelex.core.stuffs.text_content import TextContent
 from pipelex.pipeline.pipeline_models import SpecialPipelineId
+
+
+class DictPipeOutput(BaseModel):
+    working_memory: DictWorkingMemory
+    pipeline_run_id: str
 
 
 class PipeOutput(BaseModel):